Switch package names: arkd-wallet > arkd-wallet-btcwallet (to be deprecated), arkd-wallet-nbxplorer > arkd-wallet
