+++ Makefile
@@ -1,4 +1,4 @@
-.PHONY: build build-all clean cov docker-run docker-stop droppg droppgtest help integrationtest lint migrate pg pgsqlc pgtest pgmigrate psql proto proto-lint run run-light run-wallet-bitcoind run-wallet-neutrino sqlc test vet
+.PHONY: build build-all clean cov docker-run docker-stop droppg droppgtest help integrationtest lint migrate pg pgsqlc pgtest pgmigrate psql proto proto-lint run run-light run-wallet run-wallet-bitcoind run-wallet-neutrino sqlc test vet
 
 define setup_env
     $(eval include $(1))
@@ -154,14 +154,21 @@ docker-stop:
 	@echo "Stopping dockerized arkd and arkd wallet in test mode on regtest..."
 	@docker compose -f docker-compose.regtest.yml down -v
 
-## run-wallet-bitcoind: run arkd wallet in dev mode with bitcoind on regtest
+## run-wallet: run arkd wallet based on nbxplorer in dev mode on regtest
+run-wallet:
+	@echo "Running arkd wallet in dev mode with Bitcoin Core on regtest..."
+	@docker compose -f docker-compose.regtest.yml up -d pgnbxplorer nbxplorer
+	$(call setup_env, envs/arkd-wallet.regtest.env)
+	@go run ./cmd/arkd-wallet
+
+## run-wallet-bitcoind: run arkd wallet based on btcwallet in dev mode with bitcoind on regtest
 run-wallet-bitcoind:
 	@echo "Running arkd wallet in dev mode with Bitcoin Core on regtest..."
 	$(call setup_env, envs/arkd-wallet.bitcoind.regtest.env)
-	@go run ./cmd/arkd-wallet
+	@go run ./cmd/arkd-wallet-btcwallet
 
-## run-wallet-neutrino: run arkd wallet in dev mode with neutrino on regtest
+## run-wallet-neutrino: run arkd wallet based on btcwallet  in dev mode with neutrino on regtest
 run-wallet-neutrino:
 	@echo "Running arkd wallet in dev mode with Neutrino on regtest..."
 	$(call setup_env, envs/arkd-wallet.neutrino.regtest.env)
-	@go run ./cmd/arkd-wallet
+	@go run ./cmd/arkd-wallet-btcwallet
