+++ envs/arkd-wallet.regtest.env
@@ -1,0 +1,4 @@
+ARKD_WALLET_LOG_LEVEL=5
+ARKD_WALLET_NBXPLORER_URL=http://localhost:32838
+ARKD_WALLET_DATADIR=./data/regtest
+ARKD_WALLET_NETWORK=regtest
