+++ pkg/arkd-wallet-btcwallet/core/bitcoind.go
