+++ pkg/arkd-wallet-btcwallet/core/config.go
