+++ pkg/arkd-wallet-btcwallet/core/option.go
