+++ pkg/arkd-wallet-btcwallet/core/types.go
