+++ pkg/arkd-wallet-btcwallet/core/wallet_service.go
