+++ pkg/arkd-wallet-btcwallet/interface/grpc/handlers/wallet_handler.go
@@ -2,11 +2,9 @@ package handlers
 
 import (
 	"context"
-	"errors"
 
 	arkwalletv1 "github.com/arkade-os/arkd/api-spec/protobuf/gen/arkwallet/v1"
-	application "github.com/arkade-os/arkd/pkg/arkd-wallet-nbxplorer/core/application"
-	"github.com/arkade-os/arkd/pkg/arkd-wallet-nbxplorer/core/infrastructure/nbxplorer"
+	application "github.com/arkade-os/arkd/pkg/arkd-wallet-btcwallet/core"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcd/wire"
 	"google.golang.org/grpc/codes"
@@ -14,16 +12,15 @@ import (
 )
 
 type WalletServiceHandler struct {
-	wallet  application.WalletService
-	scanner application.BlockchainScanner
+	walletSvc application.WalletService
 }
 
-func NewWalletServiceHandler(walletSvc application.WalletService, scanner application.BlockchainScanner) arkwalletv1.WalletServiceServer {
-	return &WalletServiceHandler{wallet: walletSvc, scanner: scanner}
+func NewWalletServiceHandler(walletSvc application.WalletService) arkwalletv1.WalletServiceServer {
+	return &WalletServiceHandler{walletSvc: walletSvc}
 }
 
 func (h *WalletServiceHandler) GenSeed(ctx context.Context, _ *arkwalletv1.GenSeedRequest) (*arkwalletv1.GenSeedResponse, error) {
-	seed, err := h.wallet.GenSeed(ctx)
+	seed, err := h.walletSvc.GenSeed(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -31,36 +28,38 @@ func (h *WalletServiceHandler) GenSeed(ctx context.Context, _ *arkwalletv1.GenSe
 }
 
 func (h *WalletServiceHandler) Create(ctx context.Context, req *arkwalletv1.CreateRequest) (*arkwalletv1.CreateResponse, error) {
-	if err := h.wallet.Create(ctx, req.GetSeed(), req.GetPassword()); err != nil {
+	if err := h.walletSvc.Create(ctx, req.GetSeed(), req.GetPassword()); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.CreateResponse{}, nil
 }
 
 func (h *WalletServiceHandler) Restore(ctx context.Context, req *arkwalletv1.RestoreRequest) (*arkwalletv1.RestoreResponse, error) {
-	if err := h.wallet.Restore(ctx, req.GetSeed(), req.GetPassword()); err != nil {
+	if err := h.walletSvc.Restore(ctx, req.GetSeed(), req.GetPassword()); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.RestoreResponse{}, nil
 }
 
 func (h *WalletServiceHandler) Unlock(ctx context.Context, req *arkwalletv1.UnlockRequest) (*arkwalletv1.UnlockResponse, error) {
-	if err := h.wallet.Unlock(ctx, req.GetPassword()); err != nil {
+	if err := h.walletSvc.Unlock(ctx, req.GetPassword()); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.UnlockResponse{}, nil
 }
 
 func (h *WalletServiceHandler) Lock(ctx context.Context, req *arkwalletv1.LockRequest) (*arkwalletv1.LockResponse, error) {
-	if err := h.wallet.Lock(ctx); err != nil {
+	if err := h.walletSvc.Lock(ctx); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.LockResponse{}, nil
 }
 
 func (h *WalletServiceHandler) Status(ctx context.Context, _ *arkwalletv1.StatusRequest) (*arkwalletv1.StatusResponse, error) {
-	status := h.wallet.Status(ctx)
-
+	status, err := h.walletSvc.Status(ctx)
+	if err != nil {
+		return nil, err
+	}
 	return &arkwalletv1.StatusResponse{
 		Initialized: status.IsInitialized,
 		Unlocked:    status.IsUnlocked,
@@ -69,7 +68,7 @@ func (h *WalletServiceHandler) Status(ctx context.Context, _ *arkwalletv1.Status
 }
 
 func (h *WalletServiceHandler) GetPubkey(ctx context.Context, _ *arkwalletv1.GetPubkeyRequest) (*arkwalletv1.GetPubkeyResponse, error) {
-	pubkey, err := h.wallet.GetPubkey(ctx)
+	pubkey, err := h.walletSvc.GetPubkey(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -77,14 +76,14 @@ func (h *WalletServiceHandler) GetPubkey(ctx context.Context, _ *arkwalletv1.Get
 }
 
 func (h *WalletServiceHandler) GetNetwork(ctx context.Context, _ *arkwalletv1.GetNetworkRequest) (*arkwalletv1.GetNetworkResponse, error) {
-	network := h.wallet.GetNetwork(ctx)
+	network := h.walletSvc.GetNetwork(ctx)
 	return &arkwalletv1.GetNetworkResponse{Network: network}, nil
 }
 
 func (h *WalletServiceHandler) GetForfeitAddress(
 	ctx context.Context, req *arkwalletv1.GetForfeitAddressRequest,
 ) (*arkwalletv1.GetForfeitAddressResponse, error) {
-	addr, err := h.wallet.GetForfeitAddress(ctx)
+	addr, err := h.walletSvc.GetForfeitAddress(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -94,7 +93,7 @@ func (h *WalletServiceHandler) GetForfeitAddress(
 func (h *WalletServiceHandler) WatchScripts(
 	ctx context.Context, request *arkwalletv1.WatchScriptsRequest,
 ) (*arkwalletv1.WatchScriptsResponse, error) {
-	if err := h.scanner.WatchScripts(ctx, request.Scripts); err != nil {
+	if err := h.walletSvc.WatchScripts(ctx, request.Scripts); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.WatchScriptsResponse{}, nil
@@ -103,7 +102,7 @@ func (h *WalletServiceHandler) WatchScripts(
 func (h *WalletServiceHandler) UnwatchScripts(
 	ctx context.Context, request *arkwalletv1.UnwatchScriptsRequest,
 ) (*arkwalletv1.UnwatchScriptsResponse, error) {
-	if err := h.scanner.UnwatchScripts(ctx, request.Scripts); err != nil {
+	if err := h.walletSvc.UnwatchScripts(ctx, request.Scripts); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.UnwatchScriptsResponse{}, nil
@@ -112,7 +111,7 @@ func (h *WalletServiceHandler) UnwatchScripts(
 func (h *WalletServiceHandler) DeriveConnectorAddress(
 	ctx context.Context, _ *arkwalletv1.DeriveConnectorAddressRequest,
 ) (*arkwalletv1.DeriveConnectorAddressResponse, error) {
-	addr, err := h.wallet.DeriveConnectorAddress(ctx)
+	addr, err := h.walletSvc.DeriveConnectorAddress(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -122,7 +121,7 @@ func (h *WalletServiceHandler) DeriveConnectorAddress(
 func (h *WalletServiceHandler) DeriveAddresses(
 	ctx context.Context, req *arkwalletv1.DeriveAddressesRequest,
 ) (*arkwalletv1.DeriveAddressesResponse, error) {
-	addresses, err := h.wallet.DeriveAddresses(ctx, int(req.Num))
+	addresses, err := h.walletSvc.DeriveAddresses(ctx, int(req.Num))
 	if err != nil {
 		return nil, err
 	}
@@ -132,7 +131,7 @@ func (h *WalletServiceHandler) DeriveAddresses(
 func (h *WalletServiceHandler) SignTransaction(
 	ctx context.Context, req *arkwalletv1.SignTransactionRequest,
 ) (*arkwalletv1.SignTransactionResponse, error) {
-	tx, err := h.wallet.SignTransaction(ctx, req.PartialTx, req.ExtractRawTx, nil)
+	tx, err := h.walletSvc.SignTransaction(ctx, req.PartialTx, req.ExtractRawTx)
 	if err != nil {
 		return nil, err
 	}
@@ -146,7 +145,7 @@ func (h *WalletServiceHandler) SignTransactionTapscript(
 	for _, v := range req.GetInputIndexes() {
 		inIndexes = append(inIndexes, int(v))
 	}
-	tx, err := h.wallet.SignTransaction(ctx, req.GetPartialTx(), false, inIndexes)
+	tx, err := h.walletSvc.SignTransactionTapscript(ctx, req.GetPartialTx(), inIndexes)
 	if err != nil {
 		return nil, err
 	}
@@ -156,13 +155,18 @@ func (h *WalletServiceHandler) SignTransactionTapscript(
 func (h *WalletServiceHandler) SelectUtxos(
 	ctx context.Context, req *arkwalletv1.SelectUtxosRequest,
 ) (*arkwalletv1.SelectUtxosResponse, error) {
-	utxos, total, err := h.wallet.SelectUtxos(ctx, req.GetAmount(), req.GetConfirmedOnly())
+	utxos, total, err := h.walletSvc.SelectUtxos(ctx, req.GetAsset(), req.GetAmount(), req.GetConfirmedOnly())
 	if err != nil {
 		return nil, err
 	}
 	var respUtxos []*arkwalletv1.TxInput
 	for _, u := range utxos {
-		respUtxos = append(respUtxos, toTxInput(u))
+		respUtxos = append(respUtxos, &arkwalletv1.TxInput{
+			Txid:   u.Txid,
+			Index:  u.Index,
+			Script: u.Script,
+			Value:  u.Value,
+		})
 	}
 	return &arkwalletv1.SelectUtxosResponse{Utxos: respUtxos, TotalAmount: total}, nil
 }
@@ -170,7 +174,7 @@ func (h *WalletServiceHandler) SelectUtxos(
 func (h *WalletServiceHandler) BroadcastTransaction(
 	ctx context.Context, req *arkwalletv1.BroadcastTransactionRequest,
 ) (*arkwalletv1.BroadcastTransactionResponse, error) {
-	txid, err := h.wallet.BroadcastTransaction(ctx, req.GetTxs()...)
+	txid, err := h.walletSvc.BroadcastTransaction(ctx, req.GetTxs()...)
 	if err != nil {
 		return nil, err
 	}
@@ -180,7 +184,7 @@ func (h *WalletServiceHandler) BroadcastTransaction(
 func (h *WalletServiceHandler) EstimateFees(
 	ctx context.Context, req *arkwalletv1.EstimateFeesRequest,
 ) (*arkwalletv1.EstimateFeesResponse, error) {
-	fee, err := h.wallet.EstimateFees(ctx, req.GetPsbt())
+	fee, err := h.walletSvc.EstimateFees(ctx, req.GetPsbt())
 	if err != nil {
 		return nil, err
 	}
@@ -190,7 +194,7 @@ func (h *WalletServiceHandler) EstimateFees(
 func (h *WalletServiceHandler) FeeRate(
 	ctx context.Context, _ *arkwalletv1.FeeRateRequest,
 ) (*arkwalletv1.FeeRateResponse, error) {
-	feeRate, err := h.wallet.FeeRate(ctx)
+	feeRate, err := h.walletSvc.FeeRate(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -200,13 +204,18 @@ func (h *WalletServiceHandler) FeeRate(
 func (h *WalletServiceHandler) ListConnectorUtxos(
 	ctx context.Context, req *arkwalletv1.ListConnectorUtxosRequest,
 ) (*arkwalletv1.ListConnectorUtxosResponse, error) {
-	utxos, err := h.wallet.ListConnectorUtxos(ctx, req.GetConnectorAddress())
+	utxos, err := h.walletSvc.ListConnectorUtxos(ctx, req.GetConnectorAddress())
 	if err != nil {
 		return nil, err
 	}
 	respUtxos := make([]*arkwalletv1.TxInput, 0, len(utxos))
 	for _, u := range utxos {
-		respUtxos = append(respUtxos, toTxInput(u))
+		respUtxos = append(respUtxos, &arkwalletv1.TxInput{
+			Txid:   u.Txid,
+			Index:  u.Index,
+			Script: u.Script,
+			Value:  u.Value,
+		})
 	}
 	return &arkwalletv1.ListConnectorUtxosResponse{Utxos: respUtxos}, nil
 }
@@ -214,7 +223,7 @@ func (h *WalletServiceHandler) ListConnectorUtxos(
 func (h *WalletServiceHandler) MainAccountBalance(
 	ctx context.Context, _ *arkwalletv1.MainAccountBalanceRequest,
 ) (*arkwalletv1.MainAccountBalanceResponse, error) {
-	confirmed, unconfirmed, err := h.wallet.MainAccountBalance(ctx)
+	confirmed, unconfirmed, err := h.walletSvc.MainAccountBalance(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -224,7 +233,7 @@ func (h *WalletServiceHandler) MainAccountBalance(
 func (h *WalletServiceHandler) ConnectorsAccountBalance(
 	ctx context.Context, _ *arkwalletv1.ConnectorsAccountBalanceRequest,
 ) (*arkwalletv1.ConnectorsAccountBalanceResponse, error) {
-	confirmed, unconfirmed, err := h.wallet.ConnectorsAccountBalance(ctx)
+	confirmed, unconfirmed, err := h.walletSvc.ConnectorsAccountBalance(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -234,19 +243,14 @@ func (h *WalletServiceHandler) ConnectorsAccountBalance(
 func (h *WalletServiceHandler) LockConnectorUtxos(
 	ctx context.Context, req *arkwalletv1.LockConnectorUtxosRequest,
 ) (*arkwalletv1.LockConnectorUtxosResponse, error) {
-	utxos := make([]wire.OutPoint, 0, len(req.GetUtxos()))
+	utxos := make([]application.TxOutpoint, 0, len(req.GetUtxos()))
 	for _, u := range req.Utxos {
-		txhash, err := chainhash.NewHashFromStr(u.GetTxid())
-		if err != nil {
-			return nil, status.Errorf(codes.InvalidArgument, "invalid txid: %v", err)
-		}
-
-		utxos = append(utxos, wire.OutPoint{
-			Hash:  *txhash,
+		utxos = append(utxos, application.TxOutpoint{
+			Txid:  u.GetTxid(),
 			Index: u.GetIndex(),
 		})
 	}
-	if err := h.wallet.LockConnectorUtxos(ctx, utxos); err != nil {
+	if err := h.walletSvc.LockConnectorUtxos(ctx, utxos); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.LockConnectorUtxosResponse{}, nil
@@ -255,14 +259,14 @@ func (h *WalletServiceHandler) LockConnectorUtxos(
 func (h *WalletServiceHandler) GetDustAmount(
 	ctx context.Context, _ *arkwalletv1.GetDustAmountRequest,
 ) (*arkwalletv1.GetDustAmountResponse, error) {
-	dust := h.wallet.GetDustAmount(ctx)
+	dust := h.walletSvc.GetDustAmount(ctx)
 	return &arkwalletv1.GetDustAmountResponse{DustAmount: dust}, nil
 }
 
 func (h *WalletServiceHandler) GetTransaction(
 	ctx context.Context, req *arkwalletv1.GetTransactionRequest,
 ) (*arkwalletv1.GetTransactionResponse, error) {
-	tx, err := h.wallet.GetTransaction(ctx, req.GetTxid())
+	tx, err := h.walletSvc.GetTransaction(ctx, req.GetTxid())
 	if err != nil {
 		return nil, err
 	}
@@ -272,19 +276,27 @@ func (h *WalletServiceHandler) GetTransaction(
 func (h *WalletServiceHandler) SignMessage(
 	ctx context.Context, req *arkwalletv1.SignMessageRequest,
 ) (*arkwalletv1.SignMessageResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method SignMessage not implemented")
+	sig, err := h.walletSvc.SignMessage(ctx, req.GetMessage())
+	if err != nil {
+		return nil, err
+	}
+	return &arkwalletv1.SignMessageResponse{Signature: sig}, nil
 }
 
 func (h *WalletServiceHandler) VerifyMessageSignature(
 	ctx context.Context, req *arkwalletv1.VerifyMessageSignatureRequest,
 ) (*arkwalletv1.VerifyMessageSignatureResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method VerifyMessageSignature not implemented")
+	valid, err := h.walletSvc.VerifyMessageSignature(ctx, req.GetMessage(), req.GetSignature())
+	if err != nil {
+		return nil, err
+	}
+	return &arkwalletv1.VerifyMessageSignatureResponse{Valid: valid}, nil
 }
 
 func (h *WalletServiceHandler) GetCurrentBlockTime(
 	ctx context.Context, _ *arkwalletv1.GetCurrentBlockTimeRequest,
 ) (*arkwalletv1.GetCurrentBlockTimeResponse, error) {
-	ts, err := h.wallet.GetCurrentBlockTime(ctx)
+	ts, err := h.walletSvc.GetCurrentBlockTime(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -300,15 +312,8 @@ func (h *WalletServiceHandler) GetCurrentBlockTime(
 func (h *WalletServiceHandler) IsTransactionConfirmed(
 	ctx context.Context, req *arkwalletv1.IsTransactionConfirmedRequest,
 ) (*arkwalletv1.IsTransactionConfirmedResponse, error) {
-	confirmed, blocknumber, blocktime, err := h.scanner.IsTransactionConfirmed(ctx, req.GetTxid())
+	confirmed, blocknumber, blocktime, err := h.walletSvc.IsTransactionConfirmed(ctx, req.GetTxid())
 	if err != nil {
-		if errors.Is(err, nbxplorer.ErrTransactionNotFound) {
-			return &arkwalletv1.IsTransactionConfirmedResponse{
-				Confirmed:   false,
-				Blocknumber: 0,
-				Blocktime:   0,
-			}, nil
-		}
 		return nil, err
 	}
 	return &arkwalletv1.IsTransactionConfirmedResponse{
@@ -330,10 +335,10 @@ func (h *WalletServiceHandler) GetOutpointStatus(
 
 	txhash, err := chainhash.NewHashFromStr(txid)
 	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "invalid txid: %v", err)
+		return nil, err
 	}
 
-	spent, err := h.scanner.GetOutpointStatus(ctx, wire.OutPoint{
+	spent, err := h.walletSvc.GetOutpointStatus(ctx, wire.OutPoint{
 		Hash:  *txhash,
 		Index: vout,
 	})
@@ -347,7 +352,7 @@ func (h *WalletServiceHandler) GetOutpointStatus(
 func (h *WalletServiceHandler) GetReadyUpdate(
 	_ *arkwalletv1.GetReadyUpdateRequest, stream arkwalletv1.WalletService_GetReadyUpdateServer,
 ) error {
-	ch := h.wallet.GetReadyUpdate(stream.Context())
+	ch := h.walletSvc.GetReadyUpdate(stream.Context())
 	select {
 	case <-stream.Context().Done():
 		return stream.Context().Err()
@@ -363,7 +368,7 @@ func (h *WalletServiceHandler) NotificationStream(
 	_ *arkwalletv1.NotificationStreamRequest, stream arkwalletv1.WalletService_NotificationStreamServer,
 ) error {
 	ctx := stream.Context()
-	ch := h.scanner.GetNotificationChannel(ctx)
+	ch := h.walletSvc.GetNotificationChannel(ctx)
 
 	for {
 		select {
@@ -382,8 +387,8 @@ func (h *WalletServiceHandler) NotificationStream(
 				}
 				for _, v := range vtxos {
 					entry.Vtxos = append(entry.Vtxos, &arkwalletv1.VtxoWithKey{
-						Txid:  v.Txid,
-						Vout:  v.Index,
+						Txid:  v.Key.Txid,
+						Vout:  v.Key.VOut,
 						Value: v.Value,
 					})
 				}
@@ -403,19 +408,9 @@ func (h *WalletServiceHandler) NotificationStream(
 func (h *WalletServiceHandler) Withdraw(
 	ctx context.Context, req *arkwalletv1.WithdrawRequest,
 ) (*arkwalletv1.WithdrawResponse, error) {
-	txid, err := h.wallet.Withdraw(ctx, req.GetAddress(), req.GetAmount())
+	txid, err := h.walletSvc.Withdraw(ctx, req.GetAddress(), req.GetAmount())
 	if err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.WithdrawResponse{Txid: txid}, nil
-}
-
-// toTxInput converts a UTXO to a TxInput protobuf message
-func toTxInput(u application.Utxo) *arkwalletv1.TxInput {
-	return &arkwalletv1.TxInput{
-		Txid:   u.Txid,
-		Index:  u.Index,
-		Script: u.Script,
-		Value:  u.Value,
-	}
 }
