+++ pkg/arkd-wallet-btcwallet/interface/grpc/interceptors/logger.go
