+++ pkg/arkd-wallet-btcwallet/interface/grpc/interceptors/panic.go
