+++ pkg/arkd-wallet-btcwallet/interface/service.go
