+++ pkg/arkd-wallet/core/application/types.go
