+++ pkg/arkd-wallet/core/application/utils.go
