+++ pkg/arkd-wallet/core/application/wallet/key_manager.go
