+++ pkg/arkd-wallet/core/application/wallet/outpoint_locker.go
