+++ pkg/arkd-wallet/core/infrastructure/cypher/cypher.go
