+++ pkg/arkd-wallet/core/infrastructure/nbxplorer/service.go
@@ -16,7 +16,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/arkade-os/arkd/pkg/arkd-wallet-nbxplorer/core/ports"
+	"github.com/arkade-os/arkd/pkg/arkd-wallet/core/ports"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcd/wire"
 	"github.com/gorilla/websocket"
