+++ pkg/arkd-wallet/core/infrastructure/nbxplorer/types.go
