+++ pkg/arkd-wallet/core/ports/cypher.go
