+++ pkg/arkd-wallet/core/ports/seed_repository.go
