+++ pkg/arkd-wallet/interface/grpc/handlers/wallet_handler.go
@@ -2,9 +2,11 @@ package handlers
 
 import (
 	"context"
+	"errors"
 
 	arkwalletv1 "github.com/arkade-os/arkd/api-spec/protobuf/gen/arkwallet/v1"
-	application "github.com/arkade-os/arkd/pkg/arkd-wallet/core"
+	application "github.com/arkade-os/arkd/pkg/arkd-wallet/core/application"
+	"github.com/arkade-os/arkd/pkg/arkd-wallet/core/infrastructure/nbxplorer"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcd/wire"
 	"google.golang.org/grpc/codes"
@@ -12,15 +14,16 @@ import (
 )
 
 type WalletServiceHandler struct {
-	walletSvc application.WalletService
+	wallet  application.WalletService
+	scanner application.BlockchainScanner
 }
 
-func NewWalletServiceHandler(walletSvc application.WalletService) arkwalletv1.WalletServiceServer {
-	return &WalletServiceHandler{walletSvc: walletSvc}
+func NewWalletServiceHandler(walletSvc application.WalletService, scanner application.BlockchainScanner) arkwalletv1.WalletServiceServer {
+	return &WalletServiceHandler{wallet: walletSvc, scanner: scanner}
 }
 
 func (h *WalletServiceHandler) GenSeed(ctx context.Context, _ *arkwalletv1.GenSeedRequest) (*arkwalletv1.GenSeedResponse, error) {
-	seed, err := h.walletSvc.GenSeed(ctx)
+	seed, err := h.wallet.GenSeed(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -28,38 +31,36 @@ func (h *WalletServiceHandler) GenSeed(ctx context.Context, _ *arkwalletv1.GenSe
 }
 
 func (h *WalletServiceHandler) Create(ctx context.Context, req *arkwalletv1.CreateRequest) (*arkwalletv1.CreateResponse, error) {
-	if err := h.walletSvc.Create(ctx, req.GetSeed(), req.GetPassword()); err != nil {
+	if err := h.wallet.Create(ctx, req.GetSeed(), req.GetPassword()); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.CreateResponse{}, nil
 }
 
 func (h *WalletServiceHandler) Restore(ctx context.Context, req *arkwalletv1.RestoreRequest) (*arkwalletv1.RestoreResponse, error) {
-	if err := h.walletSvc.Restore(ctx, req.GetSeed(), req.GetPassword()); err != nil {
+	if err := h.wallet.Restore(ctx, req.GetSeed(), req.GetPassword()); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.RestoreResponse{}, nil
 }
 
 func (h *WalletServiceHandler) Unlock(ctx context.Context, req *arkwalletv1.UnlockRequest) (*arkwalletv1.UnlockResponse, error) {
-	if err := h.walletSvc.Unlock(ctx, req.GetPassword()); err != nil {
+	if err := h.wallet.Unlock(ctx, req.GetPassword()); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.UnlockResponse{}, nil
 }
 
 func (h *WalletServiceHandler) Lock(ctx context.Context, req *arkwalletv1.LockRequest) (*arkwalletv1.LockResponse, error) {
-	if err := h.walletSvc.Lock(ctx); err != nil {
+	if err := h.wallet.Lock(ctx); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.LockResponse{}, nil
 }
 
 func (h *WalletServiceHandler) Status(ctx context.Context, _ *arkwalletv1.StatusRequest) (*arkwalletv1.StatusResponse, error) {
-	status, err := h.walletSvc.Status(ctx)
-	if err != nil {
-		return nil, err
-	}
+	status := h.wallet.Status(ctx)
+
 	return &arkwalletv1.StatusResponse{
 		Initialized: status.IsInitialized,
 		Unlocked:    status.IsUnlocked,
@@ -68,7 +69,7 @@ func (h *WalletServiceHandler) Status(ctx context.Context, _ *arkwalletv1.Status
 }
 
 func (h *WalletServiceHandler) GetPubkey(ctx context.Context, _ *arkwalletv1.GetPubkeyRequest) (*arkwalletv1.GetPubkeyResponse, error) {
-	pubkey, err := h.walletSvc.GetPubkey(ctx)
+	pubkey, err := h.wallet.GetPubkey(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -76,14 +77,14 @@ func (h *WalletServiceHandler) GetPubkey(ctx context.Context, _ *arkwalletv1.Get
 }
 
 func (h *WalletServiceHandler) GetNetwork(ctx context.Context, _ *arkwalletv1.GetNetworkRequest) (*arkwalletv1.GetNetworkResponse, error) {
-	network := h.walletSvc.GetNetwork(ctx)
+	network := h.wallet.GetNetwork(ctx)
 	return &arkwalletv1.GetNetworkResponse{Network: network}, nil
 }
 
 func (h *WalletServiceHandler) GetForfeitAddress(
 	ctx context.Context, req *arkwalletv1.GetForfeitAddressRequest,
 ) (*arkwalletv1.GetForfeitAddressResponse, error) {
-	addr, err := h.walletSvc.GetForfeitAddress(ctx)
+	addr, err := h.wallet.GetForfeitAddress(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -93,7 +94,7 @@ func (h *WalletServiceHandler) GetForfeitAddress(
 func (h *WalletServiceHandler) WatchScripts(
 	ctx context.Context, request *arkwalletv1.WatchScriptsRequest,
 ) (*arkwalletv1.WatchScriptsResponse, error) {
-	if err := h.walletSvc.WatchScripts(ctx, request.Scripts); err != nil {
+	if err := h.scanner.WatchScripts(ctx, request.Scripts); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.WatchScriptsResponse{}, nil
@@ -102,7 +103,7 @@ func (h *WalletServiceHandler) WatchScripts(
 func (h *WalletServiceHandler) UnwatchScripts(
 	ctx context.Context, request *arkwalletv1.UnwatchScriptsRequest,
 ) (*arkwalletv1.UnwatchScriptsResponse, error) {
-	if err := h.walletSvc.UnwatchScripts(ctx, request.Scripts); err != nil {
+	if err := h.scanner.UnwatchScripts(ctx, request.Scripts); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.UnwatchScriptsResponse{}, nil
@@ -111,7 +112,7 @@ func (h *WalletServiceHandler) UnwatchScripts(
 func (h *WalletServiceHandler) DeriveConnectorAddress(
 	ctx context.Context, _ *arkwalletv1.DeriveConnectorAddressRequest,
 ) (*arkwalletv1.DeriveConnectorAddressResponse, error) {
-	addr, err := h.walletSvc.DeriveConnectorAddress(ctx)
+	addr, err := h.wallet.DeriveConnectorAddress(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -121,7 +122,7 @@ func (h *WalletServiceHandler) DeriveConnectorAddress(
 func (h *WalletServiceHandler) DeriveAddresses(
 	ctx context.Context, req *arkwalletv1.DeriveAddressesRequest,
 ) (*arkwalletv1.DeriveAddressesResponse, error) {
-	addresses, err := h.walletSvc.DeriveAddresses(ctx, int(req.Num))
+	addresses, err := h.wallet.DeriveAddresses(ctx, int(req.Num))
 	if err != nil {
 		return nil, err
 	}
@@ -131,7 +132,7 @@ func (h *WalletServiceHandler) DeriveAddresses(
 func (h *WalletServiceHandler) SignTransaction(
 	ctx context.Context, req *arkwalletv1.SignTransactionRequest,
 ) (*arkwalletv1.SignTransactionResponse, error) {
-	tx, err := h.walletSvc.SignTransaction(ctx, req.PartialTx, req.ExtractRawTx)
+	tx, err := h.wallet.SignTransaction(ctx, req.PartialTx, req.ExtractRawTx, nil)
 	if err != nil {
 		return nil, err
 	}
@@ -145,7 +146,7 @@ func (h *WalletServiceHandler) SignTransactionTapscript(
 	for _, v := range req.GetInputIndexes() {
 		inIndexes = append(inIndexes, int(v))
 	}
-	tx, err := h.walletSvc.SignTransactionTapscript(ctx, req.GetPartialTx(), inIndexes)
+	tx, err := h.wallet.SignTransaction(ctx, req.GetPartialTx(), false, inIndexes)
 	if err != nil {
 		return nil, err
 	}
@@ -155,18 +156,13 @@ func (h *WalletServiceHandler) SignTransactionTapscript(
 func (h *WalletServiceHandler) SelectUtxos(
 	ctx context.Context, req *arkwalletv1.SelectUtxosRequest,
 ) (*arkwalletv1.SelectUtxosResponse, error) {
-	utxos, total, err := h.walletSvc.SelectUtxos(ctx, req.GetAsset(), req.GetAmount(), req.GetConfirmedOnly())
+	utxos, total, err := h.wallet.SelectUtxos(ctx, req.GetAmount(), req.GetConfirmedOnly())
 	if err != nil {
 		return nil, err
 	}
 	var respUtxos []*arkwalletv1.TxInput
 	for _, u := range utxos {
-		respUtxos = append(respUtxos, &arkwalletv1.TxInput{
-			Txid:   u.Txid,
-			Index:  u.Index,
-			Script: u.Script,
-			Value:  u.Value,
-		})
+		respUtxos = append(respUtxos, toTxInput(u))
 	}
 	return &arkwalletv1.SelectUtxosResponse{Utxos: respUtxos, TotalAmount: total}, nil
 }
@@ -174,7 +170,7 @@ func (h *WalletServiceHandler) SelectUtxos(
 func (h *WalletServiceHandler) BroadcastTransaction(
 	ctx context.Context, req *arkwalletv1.BroadcastTransactionRequest,
 ) (*arkwalletv1.BroadcastTransactionResponse, error) {
-	txid, err := h.walletSvc.BroadcastTransaction(ctx, req.GetTxs()...)
+	txid, err := h.wallet.BroadcastTransaction(ctx, req.GetTxs()...)
 	if err != nil {
 		return nil, err
 	}
@@ -184,7 +180,7 @@ func (h *WalletServiceHandler) BroadcastTransaction(
 func (h *WalletServiceHandler) EstimateFees(
 	ctx context.Context, req *arkwalletv1.EstimateFeesRequest,
 ) (*arkwalletv1.EstimateFeesResponse, error) {
-	fee, err := h.walletSvc.EstimateFees(ctx, req.GetPsbt())
+	fee, err := h.wallet.EstimateFees(ctx, req.GetPsbt())
 	if err != nil {
 		return nil, err
 	}
@@ -194,7 +190,7 @@ func (h *WalletServiceHandler) EstimateFees(
 func (h *WalletServiceHandler) FeeRate(
 	ctx context.Context, _ *arkwalletv1.FeeRateRequest,
 ) (*arkwalletv1.FeeRateResponse, error) {
-	feeRate, err := h.walletSvc.FeeRate(ctx)
+	feeRate, err := h.wallet.FeeRate(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -204,18 +200,13 @@ func (h *WalletServiceHandler) FeeRate(
 func (h *WalletServiceHandler) ListConnectorUtxos(
 	ctx context.Context, req *arkwalletv1.ListConnectorUtxosRequest,
 ) (*arkwalletv1.ListConnectorUtxosResponse, error) {
-	utxos, err := h.walletSvc.ListConnectorUtxos(ctx, req.GetConnectorAddress())
+	utxos, err := h.wallet.ListConnectorUtxos(ctx, req.GetConnectorAddress())
 	if err != nil {
 		return nil, err
 	}
 	respUtxos := make([]*arkwalletv1.TxInput, 0, len(utxos))
 	for _, u := range utxos {
-		respUtxos = append(respUtxos, &arkwalletv1.TxInput{
-			Txid:   u.Txid,
-			Index:  u.Index,
-			Script: u.Script,
-			Value:  u.Value,
-		})
+		respUtxos = append(respUtxos, toTxInput(u))
 	}
 	return &arkwalletv1.ListConnectorUtxosResponse{Utxos: respUtxos}, nil
 }
@@ -223,7 +214,7 @@ func (h *WalletServiceHandler) ListConnectorUtxos(
 func (h *WalletServiceHandler) MainAccountBalance(
 	ctx context.Context, _ *arkwalletv1.MainAccountBalanceRequest,
 ) (*arkwalletv1.MainAccountBalanceResponse, error) {
-	confirmed, unconfirmed, err := h.walletSvc.MainAccountBalance(ctx)
+	confirmed, unconfirmed, err := h.wallet.MainAccountBalance(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -233,7 +224,7 @@ func (h *WalletServiceHandler) MainAccountBalance(
 func (h *WalletServiceHandler) ConnectorsAccountBalance(
 	ctx context.Context, _ *arkwalletv1.ConnectorsAccountBalanceRequest,
 ) (*arkwalletv1.ConnectorsAccountBalanceResponse, error) {
-	confirmed, unconfirmed, err := h.walletSvc.ConnectorsAccountBalance(ctx)
+	confirmed, unconfirmed, err := h.wallet.ConnectorsAccountBalance(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -243,14 +234,19 @@ func (h *WalletServiceHandler) ConnectorsAccountBalance(
 func (h *WalletServiceHandler) LockConnectorUtxos(
 	ctx context.Context, req *arkwalletv1.LockConnectorUtxosRequest,
 ) (*arkwalletv1.LockConnectorUtxosResponse, error) {
-	utxos := make([]application.TxOutpoint, 0, len(req.GetUtxos()))
+	utxos := make([]wire.OutPoint, 0, len(req.GetUtxos()))
 	for _, u := range req.Utxos {
-		utxos = append(utxos, application.TxOutpoint{
-			Txid:  u.GetTxid(),
+		txhash, err := chainhash.NewHashFromStr(u.GetTxid())
+		if err != nil {
+			return nil, status.Errorf(codes.InvalidArgument, "invalid txid: %v", err)
+		}
+
+		utxos = append(utxos, wire.OutPoint{
+			Hash:  *txhash,
 			Index: u.GetIndex(),
 		})
 	}
-	if err := h.walletSvc.LockConnectorUtxos(ctx, utxos); err != nil {
+	if err := h.wallet.LockConnectorUtxos(ctx, utxos); err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.LockConnectorUtxosResponse{}, nil
@@ -259,14 +255,14 @@ func (h *WalletServiceHandler) LockConnectorUtxos(
 func (h *WalletServiceHandler) GetDustAmount(
 	ctx context.Context, _ *arkwalletv1.GetDustAmountRequest,
 ) (*arkwalletv1.GetDustAmountResponse, error) {
-	dust := h.walletSvc.GetDustAmount(ctx)
+	dust := h.wallet.GetDustAmount(ctx)
 	return &arkwalletv1.GetDustAmountResponse{DustAmount: dust}, nil
 }
 
 func (h *WalletServiceHandler) GetTransaction(
 	ctx context.Context, req *arkwalletv1.GetTransactionRequest,
 ) (*arkwalletv1.GetTransactionResponse, error) {
-	tx, err := h.walletSvc.GetTransaction(ctx, req.GetTxid())
+	tx, err := h.wallet.GetTransaction(ctx, req.GetTxid())
 	if err != nil {
 		return nil, err
 	}
@@ -276,27 +272,19 @@ func (h *WalletServiceHandler) GetTransaction(
 func (h *WalletServiceHandler) SignMessage(
 	ctx context.Context, req *arkwalletv1.SignMessageRequest,
 ) (*arkwalletv1.SignMessageResponse, error) {
-	sig, err := h.walletSvc.SignMessage(ctx, req.GetMessage())
-	if err != nil {
-		return nil, err
-	}
-	return &arkwalletv1.SignMessageResponse{Signature: sig}, nil
+	return nil, status.Errorf(codes.Unimplemented, "method SignMessage not implemented")
 }
 
 func (h *WalletServiceHandler) VerifyMessageSignature(
 	ctx context.Context, req *arkwalletv1.VerifyMessageSignatureRequest,
 ) (*arkwalletv1.VerifyMessageSignatureResponse, error) {
-	valid, err := h.walletSvc.VerifyMessageSignature(ctx, req.GetMessage(), req.GetSignature())
-	if err != nil {
-		return nil, err
-	}
-	return &arkwalletv1.VerifyMessageSignatureResponse{Valid: valid}, nil
+	return nil, status.Errorf(codes.Unimplemented, "method VerifyMessageSignature not implemented")
 }
 
 func (h *WalletServiceHandler) GetCurrentBlockTime(
 	ctx context.Context, _ *arkwalletv1.GetCurrentBlockTimeRequest,
 ) (*arkwalletv1.GetCurrentBlockTimeResponse, error) {
-	ts, err := h.walletSvc.GetCurrentBlockTime(ctx)
+	ts, err := h.wallet.GetCurrentBlockTime(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -312,8 +300,15 @@ func (h *WalletServiceHandler) GetCurrentBlockTime(
 func (h *WalletServiceHandler) IsTransactionConfirmed(
 	ctx context.Context, req *arkwalletv1.IsTransactionConfirmedRequest,
 ) (*arkwalletv1.IsTransactionConfirmedResponse, error) {
-	confirmed, blocknumber, blocktime, err := h.walletSvc.IsTransactionConfirmed(ctx, req.GetTxid())
+	confirmed, blocknumber, blocktime, err := h.scanner.IsTransactionConfirmed(ctx, req.GetTxid())
 	if err != nil {
+		if errors.Is(err, nbxplorer.ErrTransactionNotFound) {
+			return &arkwalletv1.IsTransactionConfirmedResponse{
+				Confirmed:   false,
+				Blocknumber: 0,
+				Blocktime:   0,
+			}, nil
+		}
 		return nil, err
 	}
 	return &arkwalletv1.IsTransactionConfirmedResponse{
@@ -335,10 +330,10 @@ func (h *WalletServiceHandler) GetOutpointStatus(
 
 	txhash, err := chainhash.NewHashFromStr(txid)
 	if err != nil {
-		return nil, err
+		return nil, status.Errorf(codes.InvalidArgument, "invalid txid: %v", err)
 	}
 
-	spent, err := h.walletSvc.GetOutpointStatus(ctx, wire.OutPoint{
+	spent, err := h.scanner.GetOutpointStatus(ctx, wire.OutPoint{
 		Hash:  *txhash,
 		Index: vout,
 	})
@@ -352,7 +347,7 @@ func (h *WalletServiceHandler) GetOutpointStatus(
 func (h *WalletServiceHandler) GetReadyUpdate(
 	_ *arkwalletv1.GetReadyUpdateRequest, stream arkwalletv1.WalletService_GetReadyUpdateServer,
 ) error {
-	ch := h.walletSvc.GetReadyUpdate(stream.Context())
+	ch := h.wallet.GetReadyUpdate(stream.Context())
 	select {
 	case <-stream.Context().Done():
 		return stream.Context().Err()
@@ -368,7 +363,7 @@ func (h *WalletServiceHandler) NotificationStream(
 	_ *arkwalletv1.NotificationStreamRequest, stream arkwalletv1.WalletService_NotificationStreamServer,
 ) error {
 	ctx := stream.Context()
-	ch := h.walletSvc.GetNotificationChannel(ctx)
+	ch := h.scanner.GetNotificationChannel(ctx)
 
 	for {
 		select {
@@ -387,8 +382,8 @@ func (h *WalletServiceHandler) NotificationStream(
 				}
 				for _, v := range vtxos {
 					entry.Vtxos = append(entry.Vtxos, &arkwalletv1.VtxoWithKey{
-						Txid:  v.Key.Txid,
-						Vout:  v.Key.VOut,
+						Txid:  v.Txid,
+						Vout:  v.Index,
 						Value: v.Value,
 					})
 				}
@@ -408,9 +403,19 @@ func (h *WalletServiceHandler) NotificationStream(
 func (h *WalletServiceHandler) Withdraw(
 	ctx context.Context, req *arkwalletv1.WithdrawRequest,
 ) (*arkwalletv1.WithdrawResponse, error) {
-	txid, err := h.walletSvc.Withdraw(ctx, req.GetAddress(), req.GetAmount())
+	txid, err := h.wallet.Withdraw(ctx, req.GetAddress(), req.GetAmount())
 	if err != nil {
 		return nil, err
 	}
 	return &arkwalletv1.WithdrawResponse{Txid: txid}, nil
+}
+
+// toTxInput converts a UTXO to a TxInput protobuf message
+func toTxInput(u application.Utxo) *arkwalletv1.TxInput {
+	return &arkwalletv1.TxInput{
+		Txid:   u.Txid,
+		Index:  u.Index,
+		Script: u.Script,
+		Value:  u.Value,
+	}
 }
